allocation refactor

+++ app/Spring-Microservices/applicationservice/src/main/java/com/infinity/applicationservice/models/Allocation.java
@@ -31,9 +31,9 @@ public class Allocation {
     @Enumerated(EnumType.STRING)
     private ApplicationStatus status;
 
-    private int labPrepHours;
+    private double labPrepHours;
     private int gradingHours;
-    private int sectionHours;
+    private double sectionHours;
 
     @Column(name = "student_id")
     private Long studentId;
